Remove type coercion
