+++ vortex-array/src/aggregate_fn/erased.rs
@@ -74,11 +74,6 @@ impl AggregateFnRef {
         AggregateFnOptions { inner: &*self.0 }
     }
 
-    /// Coerce the input type for this aggregate function.
-    pub fn coerce_args(&self, input_dtype: &DType) -> VortexResult<DType> {
-        self.0.coerce_args(input_dtype)
-    }
-
     /// Compute the return [`DType`] per group given the input element type.
     ///
     /// Returns `None` if the input dtype is not supported by the aggregate function.
