+++ vortex-array/src/aggregate_fn/fns/mean/mod.rs
@@ -8,7 +8,6 @@ use crate::ArrayRef;
 use crate::ExecutionCtx;
 use crate::aggregate_fn::Accumulator;
 use crate::aggregate_fn::AggregateFnId;
-use crate::aggregate_fn::AggregateFnVTable;
 use crate::aggregate_fn::DynAccumulator;
 use crate::aggregate_fn::EmptyOptions;
 use crate::aggregate_fn::combined::BinaryCombined;
@@ -41,10 +40,8 @@ pub fn mean(array: &ArrayRef, ctx: &mut ExecutionCtx) -> VortexResult<Scalar> {
 ///
 /// Implemented as `Sum / Count` via [`BinaryCombined`].
 ///
-/// Coercion / return type:
-/// - Booleans and primitive numeric types are coerced to `f64` and the result
-///   is a nullable `f64`.
-/// - Decimals are kept as decimals but not implemented currently
+/// Booleans and primitive numeric types produce nullable `f64` results.
+/// Decimals are kept as decimals but not implemented currently.
 #[derive(Clone, Debug)]
 pub struct Mean;
 
@@ -112,35 +109,6 @@ impl BinaryCombined for Mean {
 
     fn serialize(&self, _options: &CombinedOptions<Self>) -> VortexResult<Option<Vec<u8>>> {
         unimplemented!("mean is not yet serializable");
-    }
-
-    fn coerce_args(
-        &self,
-        _options: &PairOptions<
-            <Sum as AggregateFnVTable>::Options,
-            <Count as AggregateFnVTable>::Options,
-        >,
-        input_dtype: &DType,
-    ) -> VortexResult<DType> {
-        // Advisory hint for query planners: where possible, cast input to the
-        // type we're going to compute the mean in.
-        Ok(coerced_input_dtype(input_dtype).unwrap_or_else(|| input_dtype.clone()))
-    }
-}
-
-/// Hint for callers: what to cast the input to before accumulation.
-///
-/// - Bool stays as bool — `Sum` has a native bool path and bool → f64 isn't
-///   currently a direct cast in vortex.
-/// - Primitive numerics → `f64` so the sum and finalize work without overflow.
-fn coerced_input_dtype(input_dtype: &DType) -> Option<DType> {
-    match input_dtype {
-        DType::Bool(_) => Some(input_dtype.clone()),
-        DType::Primitive(_, n) => Some(DType::Primitive(PType::F64, *n)),
-        DType::Decimal(..) => {
-            unimplemented!("mean is not implemented for decimals yet")
-        }
-        _ => None,
     }
 }
