+++ vortex-array/src/aggregate_fn/typed.rs
@@ -38,7 +38,6 @@ pub(super) trait DynAggregateFn: 'static + Send + Sync + super::sealed::Sealed {
     fn id(&self) -> AggregateFnId;
     fn options_any(&self) -> &dyn Any;
 
-    fn coerce_args(&self, input_dtype: &DType) -> VortexResult<DType>;
     fn return_dtype(&self, input_dtype: &DType) -> Option<DType>;
     fn state_dtype(&self, input_dtype: &DType) -> Option<DType>;
     fn accumulator(&self, input_dtype: &DType) -> VortexResult<AccumulatorRef>;
@@ -74,10 +73,6 @@ impl<V: AggregateFnVTable> DynAggregateFn for AggregateFnInner<V> {
 
     fn options_any(&self) -> &dyn Any {
         &self.options
-    }
-
-    fn coerce_args(&self, input_dtype: &DType) -> VortexResult<DType> {
-        V::coerce_args(&self.vtable, &self.options, input_dtype)
     }
 
     fn return_dtype(&self, input_dtype: &DType) -> Option<DType> {
