+++ vortex-array/src/expr/transform/mod.rs
@@ -2,11 +2,9 @@
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
 //! A collection of transformations that can be applied to a [`crate::expr::Expression`].
-mod coerce;
 pub(crate) mod match_between;
 mod partition;
 mod replace;
 
-pub use coerce::*;
 pub use partition::*;
 pub use replace::*;
