+++ vortex-array/src/extension/datetime/date.rs
@@ -107,29 +107,6 @@ impl ExtVTable for Date {
         Ok(())
     }
 
-    fn can_coerce_from(ext_dtype: &ExtDType<Self>, other: &DType) -> bool {
-        let DType::Extension(other_ext) = other else {
-            return false;
-        };
-        let Some(other_unit) = other_ext.metadata_opt::<Date>() else {
-            return false;
-        };
-        let our_unit = ext_dtype.metadata();
-        // We can coerce from other if our unit is finer (<=) and nullability is compatible.
-        our_unit <= other_unit && (ext_dtype.storage_dtype().is_nullable() || !other.is_nullable())
-    }
-
-    fn least_supertype(ext_dtype: &ExtDType<Self>, other: &DType) -> Option<DType> {
-        let DType::Extension(other_ext) = other else {
-            return None;
-        };
-        let other_unit = other_ext.metadata_opt::<Date>()?;
-        let our_unit = ext_dtype.metadata();
-        let finest = (*our_unit).min(*other_unit);
-        let union_null = ext_dtype.storage_dtype().nullability() | other.nullability();
-        Some(DType::Extension(Date::new(finest, union_null).erased()))
-    }
-
     fn unpack_native<'a>(
         ext_dtype: &'a ExtDType<Self>,
         storage_value: &'a ScalarValue,
@@ -188,27 +165,6 @@ mod tests {
 
         let scalar = Scalar::new(dtype, Some(ScalarValue::Primitive(PValue::I32(365))));
         assert_eq!(format!("{}", scalar.as_extension()), "1971-01-01");
-    }
-
-    #[test]
-    fn least_supertype_date_units() {
-        use crate::dtype::Nullability::NonNullable;
-
-        let days = DType::Extension(Date::new(TimeUnit::Days, NonNullable).erased());
-        let ms = DType::Extension(Date::new(TimeUnit::Milliseconds, NonNullable).erased());
-        let expected = DType::Extension(Date::new(TimeUnit::Milliseconds, NonNullable).erased());
-        assert_eq!(days.least_supertype(&ms).unwrap(), expected);
-        assert_eq!(ms.least_supertype(&days).unwrap(), expected);
-    }
-
-    #[test]
-    fn can_coerce_from_date() {
-        use crate::dtype::Nullability::NonNullable;
-
-        let days = DType::Extension(Date::new(TimeUnit::Days, NonNullable).erased());
-        let ms = DType::Extension(Date::new(TimeUnit::Milliseconds, NonNullable).erased());
-        assert!(ms.can_coerce_from(&days));
-        assert!(!days.can_coerce_from(&ms));
     }
 
     #[test]
