+++ vortex-tensor/src/types/fixed_shape_tensor/vtable.rs
@@ -33,22 +33,6 @@ impl ExtVTable for FixedShapeTensor {
         proto::deserialize(metadata)
     }
 
-    fn least_supertype(ext_dtype: &ExtDType<Self>, other: &DType) -> Option<DType> {
-        let DType::Extension(other_ext) = other else {
-            return None;
-        };
-        // Only element dtype may widen — shape, dim_names, and permutation must match exactly.
-        let other_metadata = other_ext.metadata_opt::<Self>()?;
-        if ext_dtype.metadata() != other_metadata {
-            return None;
-        }
-        let widened = ext_dtype
-            .storage_dtype()
-            .least_supertype(other_ext.storage_dtype())?;
-        let ext = ExtDType::<Self>::try_new(ext_dtype.metadata().clone(), widened).ok()?;
-        Some(DType::Extension(ext.erased()))
-    }
-
     fn validate_dtype(ext_dtype: &ExtDType<Self>) -> VortexResult<()> {
         let storage_dtype = ext_dtype.storage_dtype();
         let DType::FixedSizeList(element_dtype, list_size, _nullability) = storage_dtype else {
@@ -92,13 +76,7 @@ impl ExtVTable for FixedShapeTensor {
 
 #[cfg(test)]
 mod tests {
-    use std::sync::Arc;
-
     use rstest::rstest;
-    use vortex_array::dtype::DType;
-    use vortex_array::dtype::Nullability;
-    use vortex_array::dtype::PType;
-    use vortex_array::dtype::extension::ExtDType;
     use vortex_array::dtype::extension::ExtVTable;
     use vortex_error::VortexResult;
 
@@ -139,71 +117,5 @@ mod tests {
         #[case] metadata: VortexResult<FixedShapeTensorMetadata>,
     ) -> VortexResult<()> {
         assert_roundtrip(&metadata?)
-    }
-
-    /// Constructs a `FixedShapeTensor` ext dtype wrapped in `DType::Extension`.
-    fn tensor_dtype(
-        metadata: FixedShapeTensorMetadata,
-        element: PType,
-        list_size: u32,
-    ) -> VortexResult<DType> {
-        let storage = DType::FixedSizeList(
-            Arc::new(DType::Primitive(element, Nullability::NonNullable)),
-            list_size,
-            Nullability::NonNullable,
-        );
-        Ok(DType::Extension(
-            ExtDType::<FixedShapeTensor>::try_new(metadata, storage)?.erased(),
-        ))
-    }
-
-    #[test]
-    fn tensor_widens_element_when_metadata_matches() -> VortexResult<()> {
-        let metadata = FixedShapeTensorMetadata::new(vec![2, 3]);
-        let lhs = tensor_dtype(metadata.clone(), PType::F32, 6)?;
-        let rhs = tensor_dtype(metadata.clone(), PType::F64, 6)?;
-        let expected = tensor_dtype(metadata, PType::F64, 6)?;
-        assert_eq!(lhs.least_supertype(&rhs), Some(expected));
-        Ok(())
-    }
-
-    #[test]
-    fn tensor_different_shape_returns_none() -> VortexResult<()> {
-        let lhs = tensor_dtype(FixedShapeTensorMetadata::new(vec![2, 3]), PType::F32, 6)?;
-        let rhs = tensor_dtype(FixedShapeTensorMetadata::new(vec![3, 2]), PType::F32, 6)?;
-        assert_eq!(lhs.least_supertype(&rhs), None);
-        Ok(())
-    }
-
-    #[test]
-    fn tensor_different_permutation_returns_none() -> VortexResult<()> {
-        let lhs_metadata =
-            FixedShapeTensorMetadata::new(vec![2, 3]).with_permutation(vec![0, 1])?;
-        let rhs_metadata =
-            FixedShapeTensorMetadata::new(vec![2, 3]).with_permutation(vec![1, 0])?;
-        let lhs = tensor_dtype(lhs_metadata, PType::F32, 6)?;
-        let rhs = tensor_dtype(rhs_metadata, PType::F32, 6)?;
-        assert_eq!(lhs.least_supertype(&rhs), None);
-        Ok(())
-    }
-
-    #[test]
-    fn tensor_different_dim_names_returns_none() -> VortexResult<()> {
-        let lhs_metadata = FixedShapeTensorMetadata::new(vec![2, 3])
-            .with_dim_names(vec!["x".into(), "y".into()])?;
-        let rhs_metadata = FixedShapeTensorMetadata::new(vec![2, 3])
-            .with_dim_names(vec!["rows".into(), "cols".into()])?;
-        let lhs = tensor_dtype(lhs_metadata, PType::F32, 6)?;
-        let rhs = tensor_dtype(rhs_metadata, PType::F32, 6)?;
-        assert_eq!(lhs.least_supertype(&rhs), None);
-        Ok(())
-    }
-
-    #[test]
-    fn tensor_vs_non_extension_returns_none() -> VortexResult<()> {
-        let lhs = tensor_dtype(FixedShapeTensorMetadata::new(vec![2, 3]), PType::F32, 6)?;
-        let rhs = DType::Primitive(PType::F32, Nullability::NonNullable);
-        assert_eq!(lhs.least_supertype(&rhs), None);
-        Ok(())
     }
 }
